Misc fixes.

+++ Source/Game/Entities/Transport.cs
@@ -547,7 +547,7 @@ namespace Game.Entities
         void LoadStaticPassengers()
         {
             uint mapId = (uint)GetGoInfo().MoTransport.SpawnMap;
-            var cells = Global.ObjectMgr.GetMapObjectGuids(mapId, (byte)GetMap().GetDifficultyID());
+            var cells = Global.ObjectMgr.GetMapObjectGuids(mapId, GetMap().GetDifficultyID());
             if (cells == null)
                 return;
             foreach (var cell in cells)
